enable code-coverage for GwtMockito tests
By default jacoco java agent instruments classes on-the-fly during the test execution. This simplifies code coverage analysis a lot, as no pre-instrumentation and classpath tweaking is required.
However when using tests with custom classloader (tests that run with GwtMockitoTestRunner.class) the jacoco plugin shows zero coverage. This is because the ids of the classes in the report (jacoco.exec)
do not match the ids of the actual classes in the project (as GwtMockito uses proxies of the classes that are being tested). This problem has been solved by setting jacoco plugin to instrument
the classes before the test are run (and the proxies are created). This way exection report info is created from the original classes instead of the created proxies. The result is that the coverage of both
GwtMockito and "normal" tests can be measured. No negative side-effects (slowdown or misleading info in the report) have been observed.