Remove droolsjbpm-tools from repostory-list
The repo does not build right now, because the jbosstools eclipse target libs were moved/removed:
'No repository found at http://download.jboss.org/jbosstools/updates/requirements/webtools/R-3.6.3-20150216091848'

This branch is utterly old and unmaintained. The repo is not even part of the prod. builds, so the easiest wourkaround is to remove the repo from the build chain.